Moved verification logic to new component
