+++ src/ts/Signin/api.tsx
@@ -21,3 +21,14 @@ export function resend(data) {
   })
   .then(res => res.json());
 }
+
+export function verify(data) {
+  return fetch(API.verify, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(data),
+  })
+  .then(res => res.json());
+}
